pyproject.toml
[project]
name = "kindle-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mouseinfo==0.1.3",
  "packaging==24.1",
  "pillow==10.4.0",
  "pyautogui==0.9.54",
  "pygetwindow==0.0.9",
  "pymsgbox==1.0.9",
  "pyperclip==1.9.0",
  "pyrect==0.2.0",
  "pyscreeze==1.0.1",
  "pytesseract==0.3.13",
  "pytweening==1.2.0",
  "matplotlib==3.8.2",
  "python-dotenv>=1.0.1",
  "llama-parse>=0.1.2",
  "tqdm==4.66.2",
]

[tool.setuptools]
package-dir = { "" = "src" }
